requirements.txt
pre-commit
chalice
pytest
aws-cdk.aws-s3
aws-cdk.aws-codepipeline
aws-cdk.aws-codepipeline-actions
aws-cdk.aws-codebuild
